renamed test_seed to dbseed. this seeds csv to postgres db
must seed with knex seed:run command
